+++ src/Sass/components/_lists.scss
@@ -1,15 +1,22 @@
 @media screen and (max-width: 1023px) {
     .card-image {
         width: 80vw;
+        height: 80vw;
         display: block;
         margin-left: auto;
         margin-right: auto;
     }
 
     .cards {
-        width: 100vw;
-        text-align: center;
-        height: auto;
+        display: flex;
+        flex-wrap: wrap;
+        flex-direction: row;
+        justify-content: center;
+    }
+
+    .individual-card {
+        width: 80vw;
+        height: 80vw;
         margin: 20px;
     }
 
@@ -32,15 +39,22 @@
 
     .card-image {
         width: 300px;
+        height: 300px;
         display: block;
         margin-left: auto;
         margin-right: auto;
     }
 
     .cards {
-        text-align: center;
+        display: flex;
+        flex-wrap: wrap;
+        flex-direction: row;
+        justify-content: center;
+    }
+    
+    .individual-card {
         width: 30vw;
-        height: auto;
+        height: 30vw;
         margin: 20px;
     }
